requirements.txt
openai
python-dotenv
requests
beautifulsoup4
selenium
